Created a colission simulator for insperation to mimic CPA and TCPA

+++ V3/Model/Coordinates.cs
@@ -1,12 +1,12 @@
 public class Coordinates
 {
     public int ID { get; set; }
-    public int Y { get; set; }
-    public int X { get; set; }
+    public float Y { get; set; }
+    public float X { get; set; }
     public Symbol_flag Type;
 
     public Coordinates() { } // EF Core needs this
-    public Coordinates(Symbol_flag type, int y = 0, int x = 0)
+    public Coordinates(Symbol_flag type, float y = 0, float x = 0)
     {
         Type = type;
         Y = y;
@@ -15,9 +15,9 @@ public class Coordinates
         // Entities.Add(this);
     }
     private void Logic() => throw new NotImplementedException();
-    public void Update_Y(int y) => Y += y;
-    public void Update_X(int x) => X += x;
-    public void Update(int x, int y)
+    public void Update_Y(float y) => Y += y;
+    public void Update_X(float x) => X += x;
+    public void Update(float x,  float y)
     {
         X += x;
         Y += y;
